requirements.txt
# نظام إدارة المبيعات بالتقسيط
# Arabic Installment Sales Management System
# متطلبات التشغيل

Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
pywebview>=6.1
